run_fastapi_docker_v2_onnx/sourcecode/requirements.txt
fastapi
uvicorn
torch
diffusers
transformers
safetensors
pillow
python-multipart
accelerate
matplotlib
tritonclient[http] 
trism
onnxruntime 
requests 
httpx
opencv-python
watchfiles
opencv-python-headless
onnxruntime-tools
optimum 
torchvision
numpy
requests